Revert "Update proc.c"
This reverts commit 8c908ec4f9f4aea3acc3d9759c1cb7b4efd7dd7f.

+++ kernel/proc.c
@@ -40,7 +40,7 @@ procinit(void)
       uint64 va = KSTACK((int) (p - proc));
       kvmmap(va, (uint64)pa, PGSIZE, PTE_R | PTE_W);
       p->kstack = va;
-      p->kstack_pa = (uint64)pa;
+      p->kstack_pa = pa;
   }
   kvminithart();
 }
